Rename useCategoricalMetadataDistribution to .svelte.ts
TanStack Query v6 hooks must live in .svelte.ts files per frontend guidelines.
